# backend/Sistema_Crud_API_PIBIC/src/main/java/com/SistemaApiCrud/SistemaCrud/dto/RespostaAlunoDTO.java
package com.SistemaApiCrud.SistemaCrud.dto;

import java.time.Instant;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class RespostaAlunoDTO {

    private Long id;

    private Long idAluno;

    private Long idCaso;

    private Long idPergunta;

    private String respostaMarcada;

    private Boolean correta;

    private Instant dataResposta;
}