# backend/Sistema_Crud_API_PIBIC/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: SistemaAPI PIBIC
  version: 1.0.0
  description: |
    API REST para gerenciamento e resolucao de casos clinicos academicos.
    As operacoes de IA aceitam somente dados sinteticos ou previamente
    desidentificados e toda saida exige revisao humana antes da publicacao.
    O recurso e educacional e nao substitui decisao clinica.
  contact:
    name: Equipe SistemaAPI PIBIC
  license:
    name: Uso academico
...
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      summary: Autentica usuario e retorna token JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Token JWT gerado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "429":
          $ref: "#/components/responses/LimiteAtingido"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
...
  /auth/demonstracao/cadastro:
    post:
      summary: Cria conta temporaria de professor (login pelo e-mail) e retorna token JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CadastroDemonstracaoRequest"
      responses:
        "201":
          description: Conta criada e autenticada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "429":
          $ref: "#/components/responses/LimiteAtingido"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
  /alunos:
    get:
      summary: Lista alunos
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Alunos encontrados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaAluno"
    post:
      summary: Cadastra aluno
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlunoRequest"
      responses:
        "201":
          description: Aluno criado
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
...
  /alunos/{id}:
    get:
      summary: Busca aluno por ID
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Aluno encontrado
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    put:
      summary: Atualiza aluno
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Aluno atualizado
    delete:
      summary: Remove aluno
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          description: Aluno removido
  /alunos/{id}/casos-disponiveis:
    get:
      summary: Lista casos publicados disponiveis para o aluno
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Casos publicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaCasoClinico"
  /alunos/{id}/casos/{casoId}/completo:
    get:
      summary: Abre o caso publicado e inicia ou retoma a tentativa do aluno
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/casoId"
      responses:
        "200":
          description: Caso completo publicado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasoClinicoAluno"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
  /alunos/{id}/casos/{casoId}/responder:
    post:
      summary: Registra respostas de um aluno para um caso publicado
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/casoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResponderCasoRequest"
      responses:
        "201":
          description: Resultado calculado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultadoCaso"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
  /alunos/{id}/historico:
    get:
      summary: Busca historico de respostas do aluno
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Historico do aluno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoricoAluno"
  /alunos/{id}/desempenho:
    get:
      summary: Busca desempenho consolidado do aluno
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Desempenho do aluno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesempenhoAluno"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    casoId:
      name: casoId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    page:
      name: page
      in: query
      description: Indice da pagina, iniciando em zero.
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      description: Quantidade de elementos por pagina.
      schema:
        type: integer
        minimum: 1
        default: 20
    sort:
      name: sort
      in: query
      description: Ordenacao no formato campo,direcao.
      schema:
        type: string
        example: nome,asc
  responses:
...
    CorpoRequisicaoExcedido:
      description: O corpo da requisicao excede o limite configurado.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: urn:sistema-api-pibic:problem:corpo-requisicao-excedido
            title: Corpo da requisicao excedido
            status: 413
            detail: O corpo da requisicao excede o limite permitido
            instance: /auth/login
            timestamp: "2026-08-22T20:00:00Z"
            correlationId: 7c973582-36bc-45a7-980f-285901008ad8
            erro: O corpo da requisicao excede o limite permitido
    LimiteAtingido:
      description: Limite de tentativas, cota de IA ou concorrencia atingido.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
        Retry-After:
          description: Segundos que o cliente deve aguardar antes de tentar novamente.
          schema:
            type: integer
            format: int64
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: urn:sistema-api-pibic:problem:limite-uso-ia
            title: Limite de uso da IA
            status: 429
            detail: O limite de geracoes de IA por minuto foi atingido
            instance: /casos/42/ia/gerar
            timestamp: "2026-08-22T20:00:00Z"
            correlationId: 7c973582-36bc-45a7-980f-285901008ad8
            erro: O limite de geracoes de IA por minuto foi atingido
...
    DadosInvalidos:
      description: A requisicao possui campos, JSON, parametros ou tipos invalidos.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NaoAutenticado:
      description: Token ausente, invalido, expirado ou usuario desativado.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    NaoEncontrado:
      description: O recurso solicitado nao foi encontrado.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  headers:
    CorrelationId:
      description: Identificador da requisicao para rastreamento ponta a ponta.
      schema:
        type: string
        minLength: 1
        maxLength: 64
  schemas:
...
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
        - timestamp
        - correlationId
        - erro
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        timestamp:
          type: string
          format: date-time
        correlationId:
          type: string
        erro:
          type: string
          deprecated: true
          description: Alias legado de detail; mantido temporariamente para compatibilidade.
        errors:
          type: object
          additionalProperties:
            type: string
          description: Erros por campo quando a validacao falha.
        campos:
          type: object
          deprecated: true
          additionalProperties:
            type: string
          description: Alias legado de errors; mantido temporariamente para compatibilidade.
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          maxLength: 100
        password:
          type: string
          format: password
          maxLength: 72
...
    CadastroDemonstracaoRequest:
      type: object
      required: [nome, sobrenome, email, senha, areaAtuacao, cargo, aceiteTermos]
      properties:
        nome:
          type: string
          maxLength: 70
        sobrenome:
          type: string
          maxLength: 79
        email:
          type: string
          format: email
          maxLength: 100
          description: Em minusculas, passa a ser o usuario de acesso da conta.
        senha:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          description: Deve conter letras e numeros.
        areaAtuacao:
          type: string
          maxLength: 120
        cargo:
          type: string
          maxLength: 60
        aceiteTermos:
          type: boolean
          description: Precisa ser true.
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        tipo:
          type: string
          example: Bearer
        expiraEm:
          type: string
          format: date-time
        idUsuario:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
          enum: [ADMIN, PROFESSOR, ALUNO]
        idAluno:
          type: integer
          format: int64
          nullable: true
        idProfessor:
          type: integer
          format: int64
          nullable: true
...
    AlunoRequest:
      type: object
      required:
        - nome
        - email
        - curso
        - periodo
      properties:
        nome:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        curso:
          type: string
          maxLength: 120
        periodo:
          type: string
          maxLength: 20
    AlunoResponse:
      type: object
      properties:
        idAluno:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        curso:
          type: string
        periodo:
          type: string
...
    CasoClinicoRequest:
      type: object
      required:
        - titulo
        - disciplina
        - areaSaude
        - estilo
        - especialidade
        - nivelDificuldade
      properties:
        idProfessor:
          type: integer
          format: int64
          nullable: true
        titulo:
          type: string
          maxLength: 200
        disciplina:
          type: string
          maxLength: 120
        areaSaude:
          type: string
          maxLength: 120
        estilo:
          type: string
          maxLength: 60
        especialidade:
          type: string
          maxLength: 120
        objetivoAprendizagem:
          type: string
          maxLength: 10000
        nivelDificuldade:
          type: string
          enum: [BAIXA, MEDIA, ALTA]
        tempoLimiteMinutos:
          type: integer
          minimum: 15
          maximum: 480
          default: 60
          description: Tempo total para responder o caso completo.
    CasoClinicoResponse:
      allOf:
        - $ref: "#/components/schemas/CasoClinicoRequest"
        - type: object
          properties:
            idCaso:
              type: integer
              format: int64
            status:
              type: string
              enum: [RASCUNHO, PUBLICADO, ARQUIVADO]
            dataCriacao:
              type: string
              format: date-time
            dataAtualizacao:
              type: string
              format: date-time
...
    AlternativaAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        letra:
          type: string
          maxLength: 10
        texto:
          type: string
          maxLength: 10000
...
    PerguntaAluno:
      type: object
      description: Pergunta sem campos de correcao ou gabarito.
      properties:
        id:
          type: integer
          format: int64
        idCaso:
          type: integer
          format: int64
        texto:
          type: string
        tipo:
          type: string
          enum: [MULTIPLA_ESCOLHA, DISCURSIVA, VERDADEIRO_FALSO, DIAGNOSTICO, CONDUTA_CLINICA]
        alternativas:
          type: array
          items:
            $ref: "#/components/schemas/AlternativaAluno"
    ConteudoClinicoAluno:
      type: object
      description: Conteudo clinico sem o diagnostico esperado.
      properties:
        idConteudo:
          type: integer
          format: int64
        idCaso:
          type: integer
          format: int64
        sintomas:
          type: string
        contexto:
          type: string
        examClinico:
          type: string
        antecClinico:
          type: string
    CasoClinicoAluno:
      type: object
      description: Visao segura do caso destinada ao aluno.
      properties:
        caso:
          $ref: "#/components/schemas/CasoClinicoResponse"
        pacientes:
          type: array
          items:
            type: object
        conteudosClinicos:
          type: array
          items:
            $ref: "#/components/schemas/ConteudoClinicoAluno"
        perguntas:
          type: array
          items:
            $ref: "#/components/schemas/PerguntaAluno"
        inicioTentativa:
          type: string
          format: date-time
        prazoFinal:
          type: string
          format: date-time
        segundosRestantes:
          type: integer
          format: int64
    RespostaAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        idAluno:
          type: integer
          format: int64
        idCaso:
          type: integer
          format: int64
        idPergunta:
          type: integer
          format: int64
        respostaMarcada:
          type: string
        correta:
          type: boolean
          nullable: true
          description: |
            Resultado da correcao automatica ou da revisao humana. O valor
            null indica que a resposta ainda aguarda revisao.
        dataResposta:
          type: string
          format: date-time
...
    PaginaBase:
      type: object
      required:
        - totalElements
        - totalPages
        - number
        - size
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
...
    PaginaAluno:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/AlunoResponse"
    PaginaCasoClinico:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/CasoClinicoResponse"
...
    ResultadoCaso:
      type: object
      properties:
        idAluno:
          type: integer
          format: int64
        idCaso:
          type: integer
          format: int64
        totalRespondidas:
          type: integer
          format: int32
        totalAvaliadas:
          type: integer
          format: int32
          description: Respostas automaticas ou revisadas incluidas no calculo da nota.
        totalPendentesRevisao:
          type: integer
          format: int32
          description: Respostas discursivas ou de conduta que aguardam revisao humana.
        totalCorretas:
          type: integer
          format: int32
        nota:
          type: number
          format: double
          description: Percentual calculado somente sobre as respostas ja avaliadas.
        respostas:
          type: array
          items:
            $ref: "#/components/schemas/RespostaAluno"
    HistoricoAluno:
      type: object
      properties:
        idAluno:
          type: integer
          format: int64
        respostas:
          type: object
          description: Pagina Spring Data de respostas.
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/RespostaAluno"
            totalElements:
              type: integer
              format: int64
            totalPages:
              type: integer
              format: int32
            number:
              type: integer
              format: int32
            size:
              type: integer
              format: int32
    DesempenhoAluno:
      type: object
      properties:
        idAluno:
          type: integer
          format: int64
        totalRespostas:
          type: integer
          format: int64
        totalAvaliadas:
          type: integer
          format: int64
          description: Respostas automaticas ou revisadas incluidas no calculo do aproveitamento.
        totalPendentesRevisao:
          type: integer
          format: int64
        totalCorretas:
          type: integer
          format: int64
        aproveitamento:
          type: number
          format: double
          description: Percentual calculado somente sobre as respostas ja avaliadas.
...
    ResponderCasoRequest:
      type: object
      properties:
        respostas:
          type: array
          items:
            type: object
            properties:
              idPergunta:
                type: integer
                format: int64
              respostaMarcada:
                type: string
                maxLength: 10000